removed unneeded type cast

+++ pinterestboardwidget.php
@@ -99,7 +99,7 @@ class PinterestBoardWidget extends Module
      * @return mixed
      */
     public function displayForm() {
-        $default_lang = (int)Configuration::get('PS_LANG_DEFAULT');
+        $default_lang = Configuration::get('PS_LANG_DEFAULT');
         $helper = new HelperForm();
         $helper->module = $this;
         $helper->name_controller = $this->name;
